requirements.txt
discord.py
python-dotenv
pyperclip
win10toast
d20